+++ src/_data/environment/eleventy.js
@@ -1,3 +1,0 @@
-const eleventy = process.env.ELEVENTY_ENV
-
-module.exports = eleventy