+++ src/site/articles/articles.11tydata.js
@@ -2,6 +2,7 @@ module.exports = function () {
   return {
     contentType: 'article',
     layout: 'site/article',
-    permalink: './site/articles/{{ slug }}/index.html'
+    permalink: './site/articles/{{ slug }}/index.html',
+    url: './articles/{{ slug }}/index.html'
   }
 }